Rename + add subclasses
