+++ python/langsmith/evaluation/_runner.py
@@ -31,6 +31,7 @@ from langsmith import run_helpers as rh
 from langsmith import utils as ls_utils
 
 from langsmith.evaluation.evaluator import EvaluationResult, EvaluationResults, RunEvaluator, run_evaluator
+from langsmith.evaluation._integrations import LangChainStringEvaluator
 
 logger = logging.getLogger(__name__)
 
@@ -499,6 +500,8 @@ def _resolve_evaluators(
     for evaluator in evaluators:
         if isinstance(evaluator, RunEvaluator):
             results.append(evaluator)
+        elif isinstance(evaluator, LangChainStringEvaluator):
+            results.append(evaluator.as_run_evaluator())
         else:
             results.append(run_evaluator(evaluator))
     return results
